+++ TGIF/tgif_backend/src/main/java/group3/tgif_backend/Model/Event.java
@@ -12,7 +12,7 @@ public class Event {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Integer id; //
+    private Long id; //
 
     @Column(name = "title_en", nullable = false)
     private String titleEn; //
